MarkovBrain execution time optimization
In the vectorToBitToInt utility function a temporary container
in which the node addresses were stored in appropriate order
is replaced by iterating over the original container in
appropriate direction. Memory reallocation is thus avoided.

This function is called at every call of DeterministcGate::update().
In the codes that rely heavily on MarkovBrains' deterministic gate,
this optimization cuts execution time approximately by a quarter.

	modified:   Utilities/Utilities.h